refactor: Moving configuration as part of the datasets abstractions module
